refactor embed APIs
