+++ examples/avmnist_multimodal/model.py
@@ -25,24 +25,20 @@ def get_model(cfg, device):
 
     if cfg.MODEL.FUSION == "late":
         fusion = Concat()
-        head = MLPDecoder(
-            cfg.MODEL.MLP_IN_DIM, cfg.MODEL.MLP_HIDDEN_DIM, cfg.MODEL.OUT_DIM, include_decoder_layers=False
-        )
+        head = MLPDecoder(cfg.MODEL.MLP_IN_DIM, cfg.MODEL.MLP_HIDDEN_DIM, cfg.MODEL.OUT_DIM, use_deep_layers=False)
     elif cfg.MODEL.FUSION == "bimodal_interaction_fusion":
         fusion = BimodalInteractionFusion(
             cfg.MODEL.MULTIPLICATIVE_FUSION_IN_DIM,
             cfg.MODEL.MULTIPLICATIVE_FUSION_OUT_DIM,
             cfg.MODEL.MULTIPLICATIVE_OUTPUT,
         )
-        head = MLPDecoder(
-            cfg.MODEL.MLP_IN_DIM, cfg.MODEL.MLP_HIDDEN_DIM, cfg.MODEL.OUT_DIM, include_decoder_layers=False
-        )
+        head = MLPDecoder(cfg.MODEL.MLP_IN_DIM, cfg.MODEL.MLP_HIDDEN_DIM, cfg.MODEL.OUT_DIM, use_deep_layers=False)
     elif cfg.MODEL.FUSION == "low_rank_tensor":
         fusion = LowRankTensorFusion(
             cfg.MODEL.LOW_RANK_TENSOR_IN_DIM, cfg.MODEL.LOW_RANK_TENSOR_OUT_DIM, cfg.MODEL.LOW_RANK_TENSOR_RANK
         )
         head = MLPDecoder(
-            cfg.MODEL.MLP_LOW_RANK_IN_DIM, cfg.MODEL.MLP_HIDDEN_DIM, cfg.MODEL.OUT_DIM, include_decoder_layers=False
+            cfg.MODEL.MLP_LOW_RANK_IN_DIM, cfg.MODEL.MLP_HIDDEN_DIM, cfg.MODEL.OUT_DIM, use_deep_layers=False
         )
 
     model = MultimodalNNTrainer(encoders, fusion, head, lr=cfg.SOLVER.BASE_LR, weight_decay=cfg.SOLVER.WEIGHT_DECAY)
