+++ examples/multiomics_mogonet/model.py
@@ -3,7 +3,7 @@ from typing import List, Optional
 from torch.nn import CrossEntropyLoss
 from yacs.config import CfgNode
 
-from kale.embed.mogonet import MogonetGCN
+from kale.embed.model_lib.mogonet import MogonetGCN
 from kale.loaddata.multiomics_datasets import SparseMultiomicsDataset
 from kale.pipeline.multiomics_trainer import MultiomicsTrainer
 from kale.predict.decode import LinearClassifier, VCDN
