+++ kale/embed/attention_cnn.py
@@ -2,8 +2,10 @@ from typing import Any, Tuple, Union
 
 import torch
 import torch.nn as nn
+from torch.nn.utils.weight_norm import weight_norm
 
 from kale.embed.positional_encoding import PositionalEncoding
+from kale.predict.class_domain_nets import FCNet
 from kale.prepdata.tensor_reshape import seq_to_spatial, spatial_to_seq
 
 
@@ -148,3 +150,80 @@ class CNNTransformer(ContextCNNGeneric):
         for p in encoder.parameters():
             if p.dim() > 1:
                 nn.init.xavier_uniform_(p)
+
+
+class BANLayer(nn.Module):
+    """
+    The bilinear Attention Network (BAN) layer is designed to apply bilinear attention between two feature sets (`v` and `q`),
+    which could represent features extracted from drugs and proteins, respectively. This layer
+    enables the interaction between these two sets of features, allowing the model to learn
+    joint representations that can be used for downstream tasks like predicting drug-protein
+    interactions.
+
+    Args:
+        input_v_dim (int): Dimensionality of the first input feature set (`v`).
+        input_q_dim (int): Dimensionality of the second input feature set (`q`).
+        hidden_dim (int): Dimensionality of the hidden layer used in the bilinear attention mechanism.
+        num_out_heads (int): Number of output heads in the bilinear attention mechanism.
+        activation (str, optional): Activation function to use in the fully connected networks for `v` and `q`.
+                             Default is "ReLU".
+        dropout (float, optional): Dropout rate to apply after each layer in the fully connected networks.
+                                   Default is 0.2.
+        num_att_maps (int, optional): Number of attention maps to generate (used in pooling). Default is 3.
+    """
+
+    def __init__(
+        self, input_v_dim, input_q_dim, hidden_dim, num_out_heads, activation="ReLU", dropout=0.2, num_att_maps: int = 3
+    ):
+        super().__init__()
+
+        self.c = 32
+        self.num_att_maps = num_att_maps
+        self.input_v_dim = input_v_dim
+        self.input_q_dim = input_q_dim
+        self.hidden_dim = hidden_dim
+        self.num_out_heads = num_out_heads
+
+        self.v_net = FCNet([input_v_dim, hidden_dim * self.num_att_maps], activation=activation, dropout=dropout)
+        self.q_net = FCNet([input_q_dim, hidden_dim * self.num_att_maps], activation=activation, dropout=dropout)
+        # self.dropout = nn.Dropout(dropout[1])
+        if 1 < num_att_maps:
+            self.p_net = nn.AvgPool1d(self.num_att_maps, stride=self.num_att_maps)
+
+        if num_out_heads <= self.c:
+            self.h_mat = nn.Parameter(torch.Tensor(1, num_out_heads, 1, hidden_dim * self.num_att_maps).normal_())
+            self.h_bias = nn.Parameter(torch.Tensor(1, num_out_heads, 1, 1).normal_())
+        else:
+            self.h_net = weight_norm(nn.Linear(hidden_dim * self.num_att_maps, num_out_heads), dim=0)
+
+        self.bn = nn.BatchNorm1d(hidden_dim)
+
+    def attention_pooling(self, v, q, att_map):
+        fusion_logits = torch.einsum("bvk,bvq,bqk->bk", (v, att_map, q))
+        if 1 < self.num_att_maps:
+            fusion_logits = fusion_logits.unsqueeze(1)  # b x 1 x d
+            fusion_logits = self.p_net(fusion_logits).squeeze(1) * self.num_att_maps  # sum-pooling
+        return fusion_logits
+
+    def forward(self, v, q, softmax=False):
+        num_v = v.size(1)
+        num_q = q.size(1)
+        if self.num_out_heads <= self.c:
+            embed_v = self.v_net(v)
+            embed_q = self.q_net(q)
+            att_maps = torch.einsum("xhyk,bvk,bqk->bhvq", (self.h_mat, embed_v, embed_q)) + self.h_bias
+        else:
+            embed_v = self.v_net(v).transpose(1, 2).unsqueeze(3)
+            embed_q = self.q_net(q).transpose(1, 2).unsqueeze(2)
+            d_ = torch.matmul(embed_v, embed_q)  # b x hidden_dim x v x q
+            att_maps = self.h_net(d_.transpose(1, 2).transpose(2, 3))  # b x v x q x num_out_heads
+            att_maps = att_maps.transpose(2, 3).transpose(1, 2)  # b x num_out_heads x v x q
+        if softmax:
+            prob_mat = nn.functional.softmax(att_maps.view(-1, self.num_out_heads, num_v * num_q), 2)
+            att_maps = prob_mat.view(-1, self.num_out_heads, num_v, num_q)
+        logits = self.attention_pooling(embed_v, embed_q, att_maps[:, 0, :, :])
+        for i in range(1, self.num_out_heads):
+            logits_i = self.attention_pooling(embed_v, embed_q, att_maps[:, i, :, :])
+            logits += logits_i
+        logits = self.bn(logits)
+        return logits, att_maps
